requirements.txt
SpeechRecognition
gTTS
openai
python-dotenv
playsound
pyaudio
